vfs-btrfs: Fix build on 32 bit platforms by using long long types

Autobuild-User(master): Rusty Russell <rusty@rustcorp.com.au>
Autobuild-Date(master): Sun Mar 24 06:17:55 CET 2013 on sn-devel-104

+++ source3/modules/vfs_btrfs.c
@@ -115,11 +115,14 @@ static struct tevent_req *btrfs_copy_chunk_send(struct vfs_handle_struct *handle
 		 * cloning. Which is 4096 by default, therefore fall back to
 		 * manual read/write on failure.
 		 */
-		DEBUG(5, ("BTRFS_IOC_CLONE_RANGE failed: %s, length %lu, "
-			  "src fd: %ld off: %lu, dest fd: %d off: %lu\n",
-			  strerror(errno), cr_args.src_length,
-			  cr_args.src_fd, cr_args.src_offset,
-			  dest_fsp->fh->fd, cr_args.dest_offset));
+		DEBUG(5, ("BTRFS_IOC_CLONE_RANGE failed: %s, length %llu, "
+			  "src fd: %lld off: %llu, dest fd: %d off: %llu\n",
+			  strerror(errno),
+			  (unsigned long long)cr_args.src_length,
+			  (long long)cr_args.src_fd,
+			  (unsigned long long)cr_args.src_offset,
+			  dest_fsp->fh->fd,
+			  (unsigned long long)cr_args.dest_offset));
 		cc_state->subreq = SMB_VFS_NEXT_COPY_CHUNK_SEND(handle,
 								cc_state, ev,
 								src_fsp,
@@ -177,7 +180,8 @@ static NTSTATUS btrfs_copy_chunk_recv(struct vfs_handle_struct *handle,
 		return status;
 	}
 
-	DEBUG(10, ("server side copy chunk copied %lu\n", cc_state->copied));
+	DEBUG(10, ("server side copy chunk copied %llu\n",
+		   (unsigned long long)cc_state->copied));
 	*copied = cc_state->copied;
 	tevent_req_received(req);
 	return NT_STATUS_OK;
